Fixed current range glitch filter using invalid sample data.
The glitch filter could occasionally use one sample of invalid data during
the computation of the "pre" mean component.  The underlying cause was
that the pre mean value was computed over a FIFO that was rolling over 1
sample too late.  This injected up to one sample of undefined data.
For a length n pre value, this error occurred on roughly (n - 1) / 8
current range transitions.  Testing shows that we were lucky on
Win10 and the data was not a huge floating point number.
Added unit test and fixed.